Updated config-app shell script to skip steps based on environment checks.

+++ config-app.sh
@@ -14,6 +14,7 @@ fi
 # $1 = URL of a Git repository.
 # $2 optional = (alpine-apps|alpine-apps54)
 # $3 optional = (alpine-nginx)
+# $4 optional = (/private/etc/hosts)
 
 if [ -z "${1}" ]; then
     printf "Missing the required first parameter which should a git repository. Exiting.\n"
@@ -29,21 +30,21 @@ printf "...\n"
 
 # Allow an app container to be specified.
 APP_CONTAINER="alpine-apps"
-if [ ! -z "${2}" ]; then
+if [ -n "${2}" ]; then
     APP_CONTAINER="${2}"
 fi
 printf "Using app container: ${APP_CONTAINER}\n"
 
 # Allow an nginx container to be specified.
 NGINX_CONTAINER="alpine-nginx"
-if [ ! -z "${3}" ]; then
+if [ -n "${3}" ]; then
     NGINX_CONTAINER="${3}"
 fi
 printf "Using NGinX container: ${NGINX_CONTAINER}\n"
 
 # Check $CMP_HOST_FILE environment variable is defined.
 CMP_HOST_FILE="/private/etc/hosts"
-if [ -z "${4}" ]; then
+if [ -n "${4}" ]; then
     CMP_HOST_FILE=$4
 fi
 printf "Using hosts file: ${CMP_HOST_FILE}\n"
@@ -67,30 +68,25 @@ fi
 
 # 3. Setup with Nginx container.
 APP_NGINX_CONF_DIR="${APPS_DIR}/${APP_NAME}/web-env"
-{ # try
-    NGINX_CONF_FILE=$(find "${APP_NGINX_CONF_DIR}" -name "*-nginx.conf")
-} || { # catch
-    NGINX_CONF_FILE=''
-}
+
+NGINX_CONF_FILE=$(find "${APP_NGINX_CONF_DIR}" -name "*-nginx.conf")
 
 echo "NGINX_CONF_FILE=${NGINX_CONF_FILE}"
 
-if [ -f "${NGINX_CONF_FILE}" ]; then
+dockerNginxId=$(docker ps -aq -f "name=${NGINX_CONTAINER}")
+if [ -f "${NGINX_CONF_FILE}" ] && [ -n "${dockerNginxId}" ]; then
+    printf "NGinX container ID: ${dockerNginxId}\n"
     NGINX_NAME=$(echo "${NGINX_CONF_FILE}" | sed "s#${WEB_ENV_DIR}\(.*\)-nginx.conf\$#\1#")
 
     # Copy the Nginx config to the correct directory.
     printf "making copy: "
     cp -v "${NGINX_CONF_FILE}" "${NGINX_CONFS_DIR}/${NGINX_NAME}-nginx.conf"
 
-    printf "Debug show all Nginx configs: "
-    ls "${NGINX_CONFS_DIR}"
-
     # Make an SSL certificate for the app.
     printf "Generating a named SSL certificate for the app ${NGINX_NAME}.docker\n"
     docker exec "${NGINX_CONTAINER}" generate-named-ssl-cert.sh "${NGINX_NAME}.docker"
 
     nginxConfigFound=$(cat $CMP_HOST_FILE | grep "${NGINX_NAME}.docker")
-    # TODO: Check if the hostname entry was found in the host file.
     if [ -z "${nginxConfigFound}" ]; then
         # Add the apps domain to host PCs' hosts file
         printf "Don't forget to add '${NGINX_NAME}.docker' to the host PCs' host file.\n"
@@ -99,14 +95,15 @@ if [ -f "${NGINX_CONF_FILE}" ]; then
     # Restart the NginX services
     printf "Restarting NginX service.\n"
     docker exec "${NGINX_CONTAINER}" nginx -s reload
-
 else
-    printf "Could not find an Nginx config at the location \"${APP_NGINX_CONF_DIR}\" to copy.\n"
+    printf "Skipping NGinX configuration.\n"
 fi
 
 # 4. Copy environment variables.
+dockerAppsId=$(docker ps -aq -f "name=${APP_CONTAINER}")
 ENV_FILE="${WEB_ENV_DIR}/env-vars.txt"
-if [ -f "${ENV_FILE}" ]; then
+if [ -f "${ENV_FILE}" ] && [ -n "${dockerAppsId}" ]; then
+    printf "Apps container ID: ${dockerAppsId}\n"
     printf "Copying ${APP_NAME} environement variables over to "${APP_CONTAINER}".\n"
     E_FILE="${DIR}/${APP_NAME}-env-vars.txt"
     rm -rf "${E_FILE}"
@@ -125,6 +122,8 @@ if [ -f "${ENV_FILE}" ]; then
     docker exec -i "${APP_CONTAINER}" /root/append-vars.sh
     printf "Copied ${APP_NAME} ENV vars to "${APP_CONTAINER}".\n"
     rm "${E_FILE}"
+else
+    printf "Skipped copy of files to the apps container.\n"
 fi
 
 # 5. Run build steps if build script is present.
@@ -133,5 +132,5 @@ if [ -f "${BUILD_SCRIPT}" ]; then
     printf "Running ${APP_NAME} build script on ${APP_CONTAINER} container.\n"
     docker exec --user "root" "${APP_CONTAINER}" bash -c "source ~/.bashrc && cd /code/${APP_NAME} && ./bin/build.sh"
 else
-    printf "No build script found at the location ${BUILD_SCRIPT}.\n"
+    printf "No build script found at the location ${BUILD_SCRIPT}. Skipping.\n"
 fi
